lambda/requirements.txt
Pillow
boto3

